Make it easier to polyfill built-in extensions
This PR adds separate interfaces for constructors, similarly to how lib.es*.d.ts define ECMAScript built-ins.

This ensures that downstream typings can extend built-in constructor with static methods when polyfilling new and/or unstable features.

This was not possible with inline type declarations, because TypeScript does not allow variable declaration merging, but it is possible with a separate interface because you can declare them with the same name downstream, and it will get merged into the base type.