Simplify / fix HTMLCollection overrides
Use an extra optional type param on HTMLCollection for `namedItem` override.

This simplifies / removes quite a lot of emitter code that was necessary to maintain just for this override.

Instead, this PR is using regular backward-compatible type system capabilities.

As an additional benefit, this removes a phantom HTMLCollectionOf class that previously appeared as a valid global, but doesn't really exist in JavaScript global object.