Don't drop "conditional" from CredentialMediationRequirement.
This removal was added in c06ba3140cf9a60621747f5760d270a8008a502d
without a specific commit message. The comment says that this enum value
depends on `isConditionalMediationAvailable`, which is true, but a
number of APIs depend on feature detection yet are not removed.

This enum value is supported in Safari (as of iOS 16 & macOS 13) and
Chromium and, without this change, developers will have to cast
"conditional" whenever used.